Module places::api::matcher

Re-exports

pub use crate::match_impl::MatchBehavior;

pub use crate::match_impl::SearchBehavior;

Structs

Adaptive

OriginOrUrl

SearchParams

SearchResult

Suggestions

Enums

MatchReason
The match reason specifies why an autocomplete search result matched a query. This can be used to filter and sort matches.

Constants

ORIGIN_SQL

URL_SQL

Traits

Matcher

Functions

accept_result
Records an accepted autocomplete match, recording the query string, and chosen URL for subsequent matches.

looks_like_origin

match_url

match_with_limit

query_flat_rows_and_then_named

search_frecent
Synchronously queries all providers for autocomplete matches, then filters the matches. This isn't cancelable yet; once a search is started, it can't be interrupted, even if the user moves on (see https://github.com/mozilla/application-services/issues/265).

split_after_host_and_port

split_after_prefix