places::api::matcher
Function search_frecent

pub fn search_frecent(
    conn: &PlacesDb,
    params: SearchParams,
) -> Result<Vec<SearchResult>>

Synchronously queries all providers for autocomplete matches, then filters the matches. This isn’t cancelable yet; once a search is started, it can’t be interrupted, even if the user moves on (see https://github.com/mozilla/application-services/issues/265).

A provider can be anything that returns URL suggestions: Places history and bookmarks, synced tabs, search engine suggestions, and search keywords.