Module places::api::places_api

Structs

APIS

PlacesApi
The entry-point to the places API. This object gives access to database connections and other helpers. It enforces that only 1 write connection can exist to the database at once.

SyncConn
Wrapper around PlacesDb that automatically sets a flag (sync_conn_active) to false when finished

SyncState

Enums

ConnectionType

Constants

GLOBAL_STATE_META_KEY

Statics

ID_COUNTER