Struct places::api::places_api::PlacesApi

pub struct PlacesApi {
    db_name: PathBuf,
    write_connection: Mutex<Option<PlacesDb>>,
    sync_state: Mutex<Option<SyncState>>,
    coop_tx_lock: Arc<Mutex<()>>,
    sync_conn_active: AtomicBool,
    id: usize,
}

The entry-point to the places API. This object gives access to database connections and other helpers. It enforces that only 1 write connection can exist to the database at once.

Fields
db_name: PathBufwrite_connection: Mutex<Option<PlacesDb>>sync_state: Mutex<Option<SyncState>>coop_tx_lock: Arc<Mutex<()>>sync_conn_active: AtomicBoolid: usize
Methods

impl PlacesApi

pub fn new(db_name: impl AsRef<Path>) -> Result<Arc<Self>>

Create a new, or fetch an already open, PlacesApi backed by a file on disk.

pub fn new_memory(db_name: &str) -> Result<Arc<Self>>

Create a new, or fetch an already open, memory-based PlacesApi. You must provide a name, but you are still able to have a single writer and many reader connections to the same memory DB open.

fn new_or_existing_into(
target: &mut HashMap<PathBuf, Weak<PlacesApi>>,
db_name: PathBuf,
delete_on_fail: bool
) -> Result<Arc<Self>>

fn new_or_existing(db_name: PathBuf) -> Result<Arc<Self>>

pub fn open_connection(&self, conn_type: ConnectionType) -> Result<PlacesDb>

Open a connection to the database.

pub fn open_sync_connection(&self) -> Result<SyncConn>

pub fn close_connection(&self, connection: PlacesDb) -> Result<()>

Close a connection to the database. If the connection is the write connection, you can re-fetch it using open_connection.

fn get_disk_persisted_state(&self, conn: &PlacesDb) -> Result<Option<String>>

fn set_disk_persisted_state(
&self,
conn: &PlacesDb,
state: &Option<String>
) -> Result<()>

pub fn sync_history(
&self,
client_init: &Sync15StorageClientInit,
key_bundle: &KeyBundle
) -> Result<SyncTelemetryPing>

pub fn sync_bookmarks(
&self,
client_init: &Sync15StorageClientInit,
key_bundle: &KeyBundle
) -> Result<SyncTelemetryPing>

pub fn do_sync_one<F>(
&self,
name: &'static str,
syncer: F
) -> Result<SyncTelemetryPing> where
F: FnOnce(&SyncConn, &mut MemoryCachedState, &mut Option<String>) -> SyncResult,

pub fn sync(
&self,
client_init: &Sync15StorageClientInit,
key_bundle: &KeyBundle
) -> Result<SyncResult>

pub fn wipe_bookmarks(&self) -> Result<()>

pub fn reset_bookmarks(&self) -> Result<()>

pub fn wipe_history(&self) -> Result<()>

pub fn reset_history(&self) -> Result<()>

pub fn new_sync_conn_interrupt_handle(&self) -> Result<SqlInterruptHandle>

Get a new interrupt handle for the sync connection.