Struct places::bookmark_sync::store::BookmarksStore

pub struct BookmarksStore<'a> {
    pub db: &'a PlacesDb,
    interruptee: &'a SqlInterruptScope,
}

Fields
db: &'a PlacesDbinterruptee: &'a SqlInterruptScope
Implementations

impl<'a> BookmarksStore<'a>

pub fn new(db: &'a PlacesDb, interruptee: &'a SqlInterruptScope) -> Self

fn stage_incoming(
&self,
inbound: IncomingChangeset,
incoming_telemetry: &mut EngineIncoming
) -> Result<ServerTimestamp>

fn has_changes(&self) -> Result<bool>

fn update_local_items_in_places<'t>(
&self,
now: Timestamp,
ops: &CompletionOps<'t>
) -> Result<()>

Builds a temporary table with the merge states of all nodes in the merged tree, then updates the local tree to match the merged tree.

Conceptually, we examine the merge state of each item, and either leave the item unchanged, upload the local side, apply the remote side, or apply and then reupload the remote side with a new structure.

fn apply_remote_items(&self, now: Timestamp) -> Result<()>

fn stage_items_to_upload(
&self,
upload_items: &[UploadItem],
upload_tombstones: &[UploadTombstone]
) -> Result<()>

Stores a snapshot of all locally changed items in a temporary table for upload. This is called from within the merge transaction, to ensure that changes made during the sync don't cause us to upload inconsistent records.

Conceptually, itemsToUpload is a transient "view" of locally changed items. The local change counter is the persistent record of items that we need to upload, so, if upload is interrupted or fails, we'll stage the items again on the next sync.

fn fetch_outgoing_records(
&self,
timestamp: ServerTimestamp
) -> Result<OutgoingChangeset>

Inflates Sync records for all staged outgoing items.

fn push_synced_items(
&self,
uploaded_at: ServerTimestamp,
records_synced: Vec<SyncGuid>
) -> Result<()>

Decrements the change counter, updates the sync status, and cleans up tombstones for successfully synced items. Sync calls this method at the end of each bookmark sync.

pub(crate) fn update_frecencies(&self) -> Result<()>