Struct places::bookmark_sync::store::Merger

pub(crate) struct Merger<'a> {
    store: &'a BookmarksStore<'a>,
    remote_time: ServerTimestamp,
    local_time: Timestamp,
    external_transaction: bool,
    telem: Option<&'a mut Engine>,
}

Fields
store: &'a BookmarksStore<'a>remote_time: ServerTimestamplocal_time: Timestampexternal_transaction: booltelem: Option<&'a mut Engine>
Methods

impl<'a> Merger<'a>

pub(crate) fn new(
store: &'a BookmarksStore,
remote_time: ServerTimestamp
) -> Self

pub(crate) fn with_telemetry(
store: &'a BookmarksStore,
remote_time: ServerTimestamp,
telem: &'a mut Engine
) -> Self

pub(crate) fn set_external_transaction(&mut self, v: bool)

Prevent (or re-enable, in principal) using begin_transaction in apply().

The assumption is that if you call this, someone higher up the call_stack is managing the transaction at that point.

pub(crate) fn merge(&mut self) -> Result<()>

fn prepare(&self) -> Result<()>

Prepares synced bookmarks for merging.

fn local_row_to_item(&self, row: &Row) -> Result<(Item, Option<Content>)>

Creates a local tree item from a row in the localItems CTE.

fn remote_row_to_item(&self, row: &Row) -> Result<(Item, Option<Content>)>

Creates a remote tree item from a row in moz_bookmarks_synced.