places::db
Module db

Structs

GLOBAL_BOOKMARK_CHANGE_COUNTERS

GlobalChangeCounterTracker

An object that can tell you whether a bookmark changing operation has happened since the object was created.

PlacesDb

PlacesInitializer

PragmaGuard

A scope guard that sets a Boolean PRAGMA to a new value, and restores the inverse of the value when dropped.

SharedPlacesDb

PlacesDB that’s behind a Mutex so it can be shared between threads

Enums

Pragma

Constants

MAX_VARIABLE_NUMBER