Struct places::db::db::SharedPlacesDb

pub struct SharedPlacesDb { /* fields omitted */ }

PlacesDB that’s behind a Mutex so it can be shared between threads

Implementations

impl SharedPlacesDb

pub fn new(db: PlacesDb) -> Self

pub fn interrupt(&self)

Interrupt any current DB operation

This method doesn’t need to take the Mutex lock (which would defeat the point of interruption, since taking the lock would need to wait for the operation to complete).