Module places::ffi::error_codes

Constants

DATABASE_BUSY
The requested operation failed because the database was busy performing operations on a separate connection to the same DB.

DATABASE_CORRUPT
The requested operation failed because the store is corrupt

DATABASE_INTERRUPTED
The requested operation failed because it was interrupted

INVALID_PLACE_INFO_CANNOT_UPDATE_ROOT
CannotUpdateRoot: Attempt to modify a root in a way that is illegal, e.g. adding a child to root________, updating properties of a root, deleting a root, etc.

INVALID_PLACE_INFO_ILLEGAL_CHANGE
IllegalChange: Attempt to change a property on a bookmark node that cannot have that property. E.g. trying to edit the URL of a folder, title of a separator, etc.

INVALID_PLACE_INFO_INVALID_PARENT
InvalidParent: Attempt to add a child to a non-folder.

INVALID_PLACE_INFO_NO_ITEM
NoItem: The GUID provided does not exist.

INVALID_PLACE_INFO_URL_TOO_LONG
UrlTooLong: The provided URL cannot be inserted, as it is over the maximum URL length.

UNEXPECTED
An unexpected error occurred which likely cannot be meaningfully handled by the application.

URL_PARSE_ERROR
A URL was provided that we failed to parse