Function places::hash::hash_url

pub fn hash_url(spec: &str) -> u64

This should be identical to the “real” mozilla::places::HashURL with no prefix arg (see also hash_url_prefix for the version with one).

This returns a u64, but only the lower 48 bits should ever be set, so casting to an i64 is totally safe and lossless. If the string has no ‘:’ in it, then the returned hash will be a 32 bit hash.