Struct places::history_sync::engine::HistoryEngine

pub struct HistoryEngine<'a> {
    pub db: &'a PlacesDb,
    // some fields omitted
}

Fields
db: &'a PlacesDb
Implementations

impl<'a> HistoryEngine<'a>

pub fn new(db: &'a PlacesDb, interruptee: &'a SqlInterruptScope) -> Self

pub fn migrate_v1_global_state(db: &PlacesDb) -> Result<()>

A utility we can kill by the end of 2019 ;) Or even mid-2019? Note that this has no self - it just takes a connection. This is to ease the migration process, because this needs to be executed before bookmarks sync, otherwise the shared, persisted global state may be written by bookmarks before we’ve had a chance to migrate declined over.