Module places::history_sync::plan

Enums

IncomingPlan
This is the action we will take locally for each incoming record. For example, IncomingPlan::Delete means we will be deleting a local record and not that we will be uploading a tombstone or deleting the record itself.

Functions

apply_plan

clamp_visit_date
Clamps a history visit date between the current date and the earliest sensible date.

finish_plan

plan_incoming_record