Struct places::msg_types::BookmarkNode

pub struct BookmarkNode {
    pub node_type: Option<i32>,
    pub guid: Option<String>,
    pub date_added: Option<i64>,
    pub last_modified: Option<i64>,
    pub parent_guid: Option<String>,
    pub position: Option<u32>,
    pub title: Option<String>,
    pub url: Option<String>,
    pub child_guids: Vec<String>,
    pub child_nodes: Vec<BookmarkNode>,
    pub have_child_nodes: Option<bool>,
}

A bookmark node.

We use a single message type for bookmarks. It covers insertion, deletion, and update, and represents all three bookmark types.

This simplifies the FFI by reducing the number of types that must go across it, and retuces boilderplate, but removes some static-ish guarantees we might have otherwise.

Note that these docs comments are internal, and don’t necessarily impact the actual API we expose to Kotlin/Swift (this is particularly true around reads).

Fields
node_type: Option<i32>

The type of this bookmark, a BookmarkType (from types.rs).

This impacts which fields may be present.

It’s illegal to attempt to change this when updating a bookmark.

Note: this probably should be an enum, but prost seems to get upset about it so we’re just using int32 for now.

Note: this is node_type and not type because type is reserved in Rust.

Always returned on reads.

Required for inserts.

Not provided for updates.

guid: Option<String>

The bookmarks guid.

Always returned on reads.

Not allowed for inserts.

Required for updates (specifies which record is being changed)

date_added: Option<i64>

Creation time, in milliseconds since the unix epoch.

May not be a local timestamp, and may shift if new devices are able to provide an earlier (but still valid) timestamp.

Always returned on reads.

Ignored for insertion and update.

last_modified: Option<i64>

Last modification time, in milliseconds since the unix epoch.

Always returned on reads.

Ignored for insertion and update.

parent_guid: Option<String>

Guid of the parent record.

Returned on reads, except for reads of the bookmark root.

Required for insertion.

On updates, if provided, we treat it as a move.

Interacts with position, see its documentation below for details on how.

position: Option<u32>

Zero based index within the parent.

Not provided on reads (for now).

Allowed for insertion.

Leaving it out means ‘end of folder’.

Allowed for updates.

If parent_guid is not provided and position is, we treat this a move within the same folder.

If parent_guid and position are both provided, we treat this as a move to / within that folder, and we insert at the requested position.

If position is not provided (and parent_guid is) then it’s treated as a move to the end of that folder.

title: Option<String>

Bookmark title. Not present for type = BookmarkType::Separator.

Returned on reads if it exists.

Required when inserting folders.

url: Option<String>

Bookmark URL. Only allowed/present for type = BookmarkType::Bookmark.

Always returned on reads (for BookmarkType::Bookmark).

Required when inserting a new bookmark.

child_guids: Vec<String>

IDs of folder children, in order. Only present for type = BookmarkType::Folder.

Returned on reads (for BookmarkType::Folder).

Forbidden for insertions and updates.

Not provided if child_nodes is provided, to avoid sending more data over the FFI than necessary.

child_nodes: Vec<BookmarkNode>

Data about folder children, in order. Only present for type = BookmarkType::Folder.

For performance reasons, this only is provided if it’s requested.

have_child_nodes: Option<bool>

Returned by reads, and used to distinguish between the cases of “empty child_nodes because the API doesn’t return children” and “empty child_nodes because this folder has no children (but we’d populate them if it had them)”.

Only required because you can’t have optional repeated.

Leaving this out is equivalent to false.

Implementations

impl BookmarkNode

pub fn into_insertable(self) -> Result<InsertableItem>

Convert the protobuf bookmark into information for insertion.

impl BookmarkNode

pub fn node_type(&self) -> i32

Returns the value of node_type, or the default value if node_type is unset.

pub fn guid(&self) -> &str

Returns the value of guid, or the default value if guid is unset.

pub fn date_added(&self) -> i64

Returns the value of date_added, or the default value if date_added is unset.

pub fn last_modified(&self) -> i64

Returns the value of last_modified, or the default value if last_modified is unset.

pub fn parent_guid(&self) -> &str

Returns the value of parent_guid, or the default value if parent_guid is unset.

pub fn position(&self) -> u32

Returns the value of position, or the default value if position is unset.

pub fn title(&self) -> &str

Returns the value of title, or the default value if title is unset.

pub fn url(&self) -> &str

Returns the value of url, or the default value if url is unset.

pub fn have_child_nodes(&self) -> bool

Returns the value of have_child_nodes, or the default value if have_child_nodes is unset.