Function places::storage::bookmarks::public_node::fetch_bookmark

pub fn fetch_bookmark(
    db: &PlacesDb, 
    item_guid: &SyncGuid, 
    get_direct_children: bool
) -> Result<Option<PublicNode>>

This is similar to fetch_tree, but does not recursively fetch children of folders.

If get_direct_children is true, it will return 1 level of folder children, otherwise it returns just their guids.

It also produces the protobuf message type directly, rather than add a special variant of this bookmark type just for this function.