places::storage
Function run_maintenance_prune

pub fn run_maintenance_prune(
    conn: &PlacesDb,
    db_size_limit: u32,
    prune_limit: u32,
) -> Result<RunMaintenanceMetrics>

Run maintenance on the places DB (prune step)

The run_maintenance_*() functions are intended to be run during idle time and will take steps to clean up / shrink the database. They’re split up so that we can time each one in the Kotlin wrapper code (This is needed because we only have access to the Glean API in Kotlin and it supports a stop-watch style API, not recording specific values).

db_size_limit is the approximate storage limit in bytes. If the database is using more space than this, some older visits will be deleted to free up space. Pass in a 0 to skip this.

prune_limit is the maximum number of visits to prune if the database is over db_size_limit