Function places::storage::history::get_visited_into

pub fn get_visited_into(
    db: &PlacesDb, 
    urls_idxs: &[(usize, Url)], 
    result: &mut [bool]
) -> Result<()>

Low level api used to implement both get_visited and the FFI get_visited call. Takes a slice where we should output the results, as well as a slice of index/url pairs.

This is done so that the FFI can more easily support returning false when asked if it’s visited an invalid URL.