places::api
Module matcher

Re-exports

pub use crate::match_impl::MatchBehavior;

pub use crate::match_impl::SearchBehavior;

Structs

SearchParams

SearchResult

Functions

accept_result

Records an accepted autocomplete match, recording the query string, and chosen URL for subsequent matches.

match_url

search_frecent

Synchronously queries all providers for autocomplete matches, then filters the matches. This isn’t cancelable yet; once a search is started, it can’t be interrupted, even if the user moves on (see https://github.com/mozilla/application-services/issues/265).

split_after_host_and_port

split_after_prefix