Module places::api::places_api

Structs

PlacesApi

The entry-point to the places API. This object gives access to database connections and other helpers. It enforces that only 1 write connection can exist to the database at once.

SyncState

Enums

ConnectionType

Constants

GLOBAL_STATE_META_KEY

Functions

get_registered_sync_engine

places_api_new

For uniffi we need to expose our Arc returning constructor as a global function :( https://github.com/mozilla/uniffi-rs/pull/1063 would fix this, but got some pushback meaning we are forced into this unfortunate workaround.