Struct places::api::places_api::PlacesApi

pub struct PlacesApi { /* fields omitted */ }

The entry-point to the places API. This object gives access to database connections and other helpers. It enforces that only 1 write connection can exist to the database at once.

Implementations

impl PlacesApi

pub fn new(db_name: impl AsRef<Path>) -> Result<Arc<Self>>

Create a new, or fetch an already open, PlacesApi backed by a file on disk.

pub fn new_memory(db_name: &str) -> Result<Arc<Self>>

Create a new, or fetch an already open, memory-based PlacesApi. You must provide a name, but you are still able to have a single writer and many reader connections to the same memory DB open.

pub fn open_connection(&self, conn_type: ConnectionType) -> Result<PlacesDb>

Open a connection to the database.

pub fn open_sync_connection(&self) -> Result<SyncConn<'_>>

pub fn close_connection(&self, connection: PlacesDb) -> Result<()>

Close a connection to the database. If the connection is the write connection, you can re-fetch it using open_connection.

pub fn sync_history(
&self,
client_init: &Sync15StorageClientInit,
key_bundle: &KeyBundle
) -> Result<SyncTelemetryPing>

pub fn sync_bookmarks(
&self,
client_init: &Sync15StorageClientInit,
key_bundle: &KeyBundle
) -> Result<SyncTelemetryPing>

pub fn do_sync_one<F>(
&self,
name: &'static str,
syncer: F
) -> Result<SyncTelemetryPing> where
F: FnOnce(&SyncConn<'_>, &mut MemoryCachedState, &mut Option<String>) -> SyncResult,

pub fn sync(
&self,
client_init: &Sync15StorageClientInit,
key_bundle: &KeyBundle
) -> Result<SyncResult>

pub fn wipe_bookmarks(&self) -> Result<()>

pub fn reset_bookmarks(&self) -> Result<()>

pub fn wipe_history(&self) -> Result<()>

pub fn reset_history(&self) -> Result<()>

pub fn new_sync_conn_interrupt_handle(&self) -> Result<SqlInterruptHandle>

Get a new interrupt handle for the sync connection.