Struct places::db::PlacesTransaction

pub struct PlacesTransaction<'conn>(_);

High level transaction type which "does the right thing" for you. Construct one with PlacesDb::begin_transaction().

Implementations

impl<'conn> PlacesTransaction<'conn>

pub fn should_commit(&self) -> bool

Returns true if the current transaction should be committed at the earliest opportunity.

pub fn maybe_commit(&mut self) -> Result<()>

For transactions on sync connnections: Checks to see if we have held a transaction for longer than the requested time, and if so, commits the current transaction and opens another.

For transactions on other connections: debug_assert!s, or logs a warning and does nothing.

pub fn commit(self) -> Result<()>

Consumes and commits a PlacesTransaction transaction.

pub fn rollback(self) -> Result<()>

Consumes and attempst to roll back a PlacesTransaction. Note that if maybe_commit has been called, this may only roll back as far as that call.