places::storage::bookmarks

Module fetch

Structs

BookmarkData

Structs we return when reading bookmarks

Folder

RECENT_BOOKMARKS_QUERY

SEARCH_QUERY

Separator

Enums

Item

Functions

fetch_bookmark

This is similar to fetch_tree, but does not recursively fetch children of folders.

fetch_bookmarks_by_url

fetch_tree

Call fetch_tree_with_depth with FetchDepth::Deepest. This is the function called by the FFI when requesting the tree.

fetch_tree_with_depth

Call fetch_tree with a depth parameter and convert the result to an Item.

recent_bookmarks

search_bookmarks